requirements/base.txt
pytz==2020.4  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.1.0  # https://github.com/celery/django-celery-beat
flower==0.9.5  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.0.11  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.43.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.10.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-timezone-field==4.0  # Fixes error with celery beat and timezones in db models

# Chart.js
django-chartjs==2.2.1  # https://github.com/peopledoc/django-chartjs

# Git Library
GitPython==3.1.26  # https://github.com/gitpython-developers/GitPython
pandas==1.4.2  # https://pypi.org/project/pandas/
numpy==1.22.3  # https://pypi.org/project/numpy/
matplotlib==3.5.2  # https://pypi.org/project/matplotlib/
seaborn==0.11.2  # https://pypi.org/project/seaborn/
scikit-learn==1.0.2  # https://pypi.org/project/scikit-learn/
scipy==1.8.0  # https://pypi.org/project/scipy/
lmfit==1.0.3  # https://pypi.org/project/lmfit/
uncertainties==3.1.6  # https://pypi.org/project/uncertainties/
lightgbm==3.3.2  # https://pypi.org/project/lightgbm/
xgboost==1.6.0  # https://pypi.org/project/xgboost/
keras==2.9.0  # https://pypi.org/project/keras/
tensorflow==2.9.0  # https://pypi.org/project/tensorflow/
